[hw4] Implement simple rotation
Now, a user can click and drag using the left mouse button to rotate the
objects. The algorithm is currently very simple, but is also intuitive.
The user simply drags the mouse in the image plane, and the angle of
rotation is proportional to the length of the drag line.