[hw6] Add animated water-like surface
Adds an animated water-like surface to the viewer.
This surface reflects an image, `sky.png`, using OpenGL's "sphere
mapping" code. The surface has "ripples" that resemble water waves,
as they are circular and propagate outward with time. The normals for
each coordinate in the water-like surface determine the reflected
texture coordinates. For specifics on the formula, see the code comments.